add styles for map[C
