+++ src/js/project/index.js
@@ -1,6 +1,5 @@
 import './sliders/_news-slider.js';
-// import './map/_inter-map.js';
-import './map/_test-map.js';
+import './map/_inter-map.js';
 import {stream} from './_stream.js'
 stream();
 import { animateHeroImage } from './_hero-animation.js'
